feat: include testing for basic example
