+++ examples/basic/package.json
@@ -9,7 +9,7 @@
   },
   "dependencies": {
     "@opentelemetry/api": "^1.0.2",
-    "@opentelemetry/instrumentation-http": "^0.23.0",
+    "@opentelemetry/instrumentation-http": "^0.24.0",
     "@splunk/otel": "^0.12.0",
     "axios": "^0.21.1",
     "express": "^4.17.1"
