+++ test/examples/basic.docker-compose.yml
@@ -1,0 +1,36 @@
+version: '3'
+services:
+  collector:
+    # hopefully 0.27.0 will include polling exporter
+    # image: quay.io/signalfx/splunk-otel-collector:0.27.0
+    image: ghcr.io/splunk-o11y-gdi-bot/splunk-otel-collector-dev:polling
+    environment:
+      - SPLUNK_CONFIG=/etc/otel/config.yml
+    volumes:
+      - ./collector.config.yml:/etc/otel/config.yml
+    # ports:
+    #   - "55681:55681" # HTTP OTLP
+    #   - "4317:4317" # GRPC OTLP
+    #   - "8378:8378" # POLLING
+    #   - "13133:13133" # HEALTH CHECK
+  app:
+    build:
+      context:  ../../examples/basic
+      dockerfile: ../../test/examples/Dockerfile
+    command: npm start
+    environment:
+      - OTEL_EXPORTER_OTLP_ENDPOINT=http://collector:55681/v1/traces
+      - PORT=80
+    volumes:
+      - ./express/app.env:/home/node/app/.env
+  test:
+    build: .
+    command: npx nodemon ./basic # for local iteration
+    volumes:
+      - ./:/home/node/app # for local iteration
+    environment:
+      - REQ_URL=http://app/hello
+      - COLLECTOR_URL=http://collector:8378
+    depends_on:
+      - app
+      - collector
