+++ test/examples/basic/app.env
@@ -1,0 +1,4 @@
+PORT=80
+OTEL_SERVICE_NAME='express-example'
+OTEL_RESOURCE_ATTRIBUTES='deployment.environment=dev'
+OTEL_LOG_LEVEL='DEBUG'
