+++ test/examples/package.json
@@ -1,16 +1,16 @@
 {
-  "name": "integration",
+  "name": "test-express-example",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
   "scripts": {
-    "start": "node .",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node ./express",
+    "test": "doco up --build --exit-code-from test"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "got": "^11.8.2",
     "nodemon": "^2.0.12"
   }
 }
